requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
aiofiles==23.2.1
pymediainfo==6.1.0
APScheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-levenshtein==0.25.0
fuzzywuzzy==0.18.0
redis==5.0.1
aioredis==2.0.1